feat: Define RecipeItemDTO and mapping to RecipeItem entity
This commit introduces the `RecipeItemDTO` data class, which represents the structure of recipe data received from an external source. It's marked as `@Serializable` for kotlinx.serialization.

Additionally, an extension function `toRecipe()` is added to `RecipeItemDTO`. This function maps the DTO to a `RecipeItem` domain entity. It handles potential null values in the DTO and constructs a list of ingredients and instructions. If the `idMeal` is null, the function returns null.

A utility extension function `capitalizeFirstWord()` for `String` is also included.